refactor: components 폴더에 index 생성, import 문 정리
