+++ src/components/CardList/index.jsx
@@ -1,7 +1,7 @@
 import Card from "components/Card";
 import styles from "./styles.module.css";
 
-export default function CardList({ links }) {
+export function CardList({ links }) {
   return (
     <ul className={styles["l_row"]}>
       {/* 옵셔널 체이닝: 이걸 이용하지 않으면 새로고침시 에러가 발생한다 */}
