+++ src/components/index.jsx
@@ -1,0 +1,5 @@
+import { Header } from "./Header";
+import { CardList } from "./CardList";
+import { Footer } from "./Footer";
+
+export { Header, CardList, Footer };
