+++ src/pages/Folder/components/index.jsx
@@ -1,0 +1,6 @@
+import { FolderAddLinkArea } from "./FolderAddLinkArea";
+import { FolderCategory } from "./FolderCategory";
+import { FolderControl } from "./FolderControl";
+import { FolderHeader } from "./FolderHeader";
+
+export { FolderAddLinkArea, FolderCategory, FolderControl, FolderHeader };
