+++ src/pages/Share/index.jsx
@@ -1,6 +1,7 @@
-import Header from "components/Header";
-import Footer from "components/Footer";
-import CardList from "components/CardList";
+// import Header from "components/Header";
+// import Footer from "components/Footer";
+// import CardList from "components/CardList";
+import { Header, Footer, CardList } from "components";
 import styles from "./styles.module.css";
 import SearchIcon from "assets/Search.png";
 import { useEffect, useState } from "react";
